Fixed hanging test issue
Mocha was being called with (ill-advised) --exit flag to exit
after testing as mongoose and/or express connections were hanging
around. HTTP server returned by app.listen() exposed as module
and added in after hooks as necessary.

This is extremely hacky and Supertest does not appear to be
actively maintained as of now. May need to consider switching to
another library for integration testing express.